Format python code with black
Formatting is done with black version 21.12.b0, because it was
the last black version supporting python2 code.

It is installed with
`pip install 'black[python2]==21.12b0'`
and
`pip install --upgrade click==8.0.2`.
Later is needed because of know uncompliance
between black 21 and newer click version.

Also *.py.example files are formatted.