Remove DIY "migration" script
web/update_db.* scripts used several non-worging commands:

* `python manage.py reset` non-working since Django 1.5
* `python tupa/dia2django.py` script removed in b15a827
* `python legacy/RenameFixture.py` didn't
     do anything in current version.
* `python manage.py inspectlegacy` Broken in code
     formatting b215abbc7ae11614bab50fb7c3b95c307cf933d5

inspectlegacy.py is originally vendored from Django's
inspectdb subcommand. It was altered to write changes
directly to legacy/models.py and to do it only for
models already specified in that same legacy/models.py.

Commit f316ada666454c30494bd7afea7f3df8975dd4a7 already changed
instructions so that those database updation scripts are not used.

In Django 1.7 there will be real migrations baked in to the framewrok.